Perf/bpe merge review (#2297)
* chore(node): sync Cargo.lock with tk-encode's libc dependency

* refactor(bpe): flatten the filename prefixes, group the released-crate BPE under legacy/

* refactor(bpe): split the tables file, document the packed merge value once

* refactor(bpe): move code next to its users

* refactor(bpe): one name per concept

* refactor(bpe): collapse the three copies of the char-as-bytes push
