+++ tokenizers/tk-encode/src/models/bpe/legacy/word.rs
@@ -1,8 +1,50 @@
-use super::Pair;
+use crate::models::bpe::Pair;
 use ahash::AHashMap;
 use dary_heap::QuaternaryHeap;
 use rand::{Rng, rng};
 use std::cmp::Ordering;
+use std::{iter, mem};
+
+/// Provides access to the `FirstLastIterator` to any Iterator
+pub trait WithFirstLastIterator: Iterator + Sized {
+    fn with_first_and_last(self) -> FirstLastIterator<Self>;
+}
+
+impl<I> WithFirstLastIterator for I
+where
+    I: Iterator,
+{
+    fn with_first_and_last(self) -> FirstLastIterator<Self> {
+        FirstLastIterator {
+            first: true,
+            iter: self.peekable(),
+        }
+    }
+}
+
+/// Provides information about whether an item is the first and/or the last of the iterator
+pub struct FirstLastIterator<I>
+where
+    I: Iterator,
+{
+    first: bool,
+    iter: iter::Peekable<I>,
+}
+
+impl<I> Iterator for FirstLastIterator<I>
+where
+    I: Iterator,
+{
+    /// (is_first, is_last, item)
+    type Item = (bool, bool, I::Item);
+
+    fn next(&mut self) -> Option<Self::Item> {
+        let first = mem::replace(&mut self.first, false);
+        self.iter
+            .next()
+            .map(|e| (first, self.iter.peek().is_none(), e))
+    }
+}
 
 #[derive(Debug, Eq)]
 pub struct Merge {
